+++ src/fasttext/args.h
@@ -7,8 +7,7 @@
  * of patent rights can be found in the PATENTS file in the same directory.
  */
 
-#ifndef FASTTEXT_ARGS_H
-#define FASTTEXT_ARGS_H
+#pragma once
 
 #include <istream>
 #include <ostream>
@@ -17,17 +16,17 @@
 
 namespace fasttext {
 
-enum class model_name : int {cbow=1, sg, sup};
-enum class loss_name : int {hs=1, ns, softmax};
+enum class model_name : int { cbow = 1, sg, sup };
+enum class loss_name : int { hs = 1, ns, softmax };
 
 class Args {
-  private:
+  protected:
     std::string lossToString(loss_name);
+    std::string boolToString(bool);
 
   public:
     Args();
     std::string input;
-    std::string test;
     std::string output;
     double lr;
     int lrUpdateRate;
@@ -48,24 +47,21 @@ class Args {
     std::string label;
     int verbose;
     std::string pretrainedVectors;
-    int saveOutput;
+    bool saveOutput;
 
-    bool qout;
-    bool retrain;
-    bool qnorm;
-    size_t cutoff;
-    size_t dsub;
+  bool qout;
+  bool retrain;
+  bool qnorm;
+  size_t cutoff;
+  size_t dsub;
 
-    void parseArgs(const std::vector<std::string>& args);
-    void printHelp();
-    void printBasicHelp();
-    void printDictionaryHelp();
-    void printTrainingHelp();
-    void printQuantizationHelp();
-    void save(std::ostream&);
-    void load(std::istream&);
+  void parseArgs(const std::vector<std::string>& args);
+  void printHelp();
+  void printBasicHelp();
+  void printDictionaryHelp();
+  void printTrainingHelp();
+  void printQuantizationHelp();
+  void save(std::ostream&);
+  void load(std::istream&);
 };
-
 }
-
-#endif