+++ src/fasttext/dictionary.cc
@@ -16,6 +16,7 @@
 #include <algorithm>
 #include <iterator>
 #include <cmath>
+#include <stdexcept>
 
 namespace fasttext {
 
@@ -92,9 +93,6 @@ void Dictionary::getSubwords(const std::string& word,
   if (i >= 0) {
     ngrams.push_back(i);
     substrings.push_back(words_[i].word);
-  } else {
-    ngrams.push_back(-1);
-    substrings.push_back(word);
   }
   computeSubwords(BOW + word + EOW, ngrams, substrings);
 }
@@ -239,9 +237,8 @@ void Dictionary::readFromFile(std::istream& in) {
     std::cerr << "Number of labels: " << nlabels_ << std::endl;
   }
   if (size_ == 0) {
-    std::cerr << "Empty vocabulary. Try a smaller -minCount value."
-              << std::endl;
-    exit(EXIT_FAILURE);
+    throw std::invalid_argument(
+        "Empty vocabulary. Try a smaller -minCount value.");
   }
 }
 
@@ -382,8 +379,10 @@ void Dictionary::pushHash(std::vector<int32_t>& hashes, int32_t id) const {
 }
 
 std::string Dictionary::getLabel(int32_t lid) const {
-  assert(lid >= 0);
-  assert(lid < nlabels_);
+  if (lid < 0 || lid >= nlabels_) {
+    throw std::invalid_argument(
+        "Label id is out of range [0, " + std::to_string(nlabels_) + "]");
+  }
   return words_[lid + nwords_].word;
 }