+++ src/fasttext/utils.h
@@ -7,10 +7,17 @@
  * of patent rights can be found in the PATENTS file in the same directory.
  */
 
-#ifndef FASTTEXT_UTILS_H
-#define FASTTEXT_UTILS_H
+#pragma once
 
 #include <fstream>
+
+#if defined(__clang__) || defined(__GNUC__)
+# define FASTTEXT_DEPRECATED(msg) __attribute__((__deprecated__(msg)))
+#elif defined(_MSC_VER)
+# define FASTTEXT_DEPRECATED(msg) __declspec(deprecated(msg))
+#else
+# define FASTTEXT_DEPRECATED(msg)
+#endif
 
 namespace fasttext {
 
@@ -21,5 +28,3 @@ namespace utils {
 }
 
 }
-
-#endif