+++ NEWS.md
@@ -1,3 +1,10 @@
+# 0.2.4
+
+  * major refactoring
+  * use last version of fastText sourcecode
+  * sentence representation function
+  * add tags function
+
 # 0.2.3 (9.11.17)
 
   * fix a cran note related to the DESCRIPTION file